Add support for moderator only message.
